Move handler functions in Product component to helper file
This is so that a code reviewer will not have to deal with so much
unnecessary details.
